Extract support config IssueParams to jira api interfaces
